+++ PexelsWiki/PexelsWiki/Domain/Ports/SpecificVideoPort.swift
@@ -1,0 +1,16 @@
+//
+//  SpecificVideoPort.swift
+//  PexelsWiki
+//
+//  Copyright (c) 2024 Jeremy All rights reserved.
+    
+
+import Foundation
+
+protocol SpecificVideoPort {
+    @discardableResult
+    func fetchVideoForID(
+        _ id: Int,
+        _ completion: @escaping (Result<SpecificVideo, Error>) -> Void
+    ) -> Cancellable?
+}
