+++ PexelsWiki/PexelsWiki/Domain/Ports/SpeficicPhotoPort.swift
@@ -1,0 +1,14 @@
+//
+//  SpeficicPhotoPort.swift
+//  PexelsWiki
+//
+//  Copyright (c) 2024 Jeremy All rights reserved.
+    
+
+protocol SpecificPhotoPort {
+    @discardableResult
+    func fetchPhotoForID(
+        _ id: Int,
+        _ completion: @escaping (Result<SpecificPhoto, Error>) -> Void
+    ) -> Cancellable?
+}
